New Doxygen Test 1

+++ int_input_checked.h
@@ -14,41 +14,41 @@
 
 using namespace std;
 
-// Function to stop input failure in its tracks (int version)
-// Parameters:
-//  * prompt: string - the prompt to display to the user
-//  * rangeBegin: int - the beginning of the range of acceptable values
-//  * rangeEnd: int - the end of the range of acceptable values
-// Returns:
-//  * int - the user's input
+/// Function to stop input failure in its tracks (int version)
+///
+/// Checks integers to make sure their values are valid. If they are not, the user gets an error and is asked to give a new input
+///@param prompt: string - the prompt to display to the user
+///@param rangeBegin: int - the beginning of the range of acceptable values
+///@param rangeEnd: int - the end of the range of acceptable values
+///@returns int - the user's input
 int intInputChecked(string prompt, int rangeBegin, int rangeEnd);
 
-// Function to stop input failure in its tracks (int version)
-// Parameters:
-//  * prompt: string - the prompt to display to the user
-//  * rangeBegin: int - the beginning of the range of acceptable values
-//  * rangeEnd: int - the end of the range of acceptable values
-// Returns:
-//  * int - the user's input
-long long intInputCheckedLong(string prompt, int rangeBegin, int rangeEnd);
+/// Function to stop input failure in its tracks (long long version)
+///
+/// Checks long long integers to make sure the input is valid. If they are not, the user gets an error and is asked to give a new input
+///@param prompt: string - the prompt to display to the user
+///@param rangeBegin: long long - the beginning of the range of acceptable values
+///@param rangeEnd: long long - the end of the range of acceptable values
+///@returns long long - the user's input
+long long intInputCheckedLong(string prompt, long long rangeBegin, long long rangeEnd);
 
-// Function to stop input failure in its tracks (float version)
-// Parameters:
-//  * prompt: string - the prompt to display to the user
-//  * rangeBegin: float - the beginning of the range of acceptable values
-//  * rangeEnd: float - the end of the range of acceptable values
-// Returns:
-//  * float - the user's input
+/// Function to stop input failure in its tracks (float version)
+///
+/// Checks floating point values to make sure the input is valid. If they are not, the user gets an error and is asked to give a new input
+///@param prompt: string - the prompt to display to the user
+///@param rangeBegin: float - the beginning of the range of acceptable values
+///@param rangeEnd: float - the end of the range of acceptable values
+///@returns float - the user's input
 float floatInputChecked(string prompt, float rangeBegin, float rangeEnd);
 
 
-// Function to stop input failure in its tracks (double version)
-// Parameters:
-//  * prompt: string - the prompt to display to the user
-//  * rangeBegin: double - the beginning of the range of acceptable values
-//  * rangeEnd: double - the end of the range of acceptable values
-// Returns:
-//  * double - the user's input
+/// Function to stop input failure in its tracks (double version)
+///
+/// Checks doubles to make sure their values are valid. If they are not, the user gets an error and is asked to give a new input
+///@param prompt: string - the prompt to display to the user
+///@param rangeBegin: double - the beginning of the range of acceptable values
+///@param rangeEnd: double - the end of the range of acceptable values
+///@returns double - the user's input
 double doubleInputChecked(string prompt, double rangeBegin, double rangeEnd);
 
 #endif
